scrap: simplyfing config
