+++ include/sniff/minimize.h
@@ -4,9 +4,14 @@
 #include <vector>
 
 #include "sniff/kmer.h"
-#include "sniff/minimize_config.h"
 
 namespace sniff {
+
+struct MinimizeConfig {
+  std::uint32_t kmer_len = 15;
+  std::uint32_t window_len = 5;
+  bool minhash = false;
+};
 
 auto Minimize(MinimizeConfig cfg, std::string_view sequence)
     -> std::vector<KMer>;
