--- include/sniff/minimize_config.h
@@ -1,13 +1,0 @@
-#pragma once
-
-#include <cstdint>
-
-namespace sniff {
-
-struct MinimizeConfig {
-  std::uint32_t kmer_len = 15;
-  std::uint32_t window_len = 5;
-  bool minhash = false;
-};
-
-}  // namespace sniff
