--- src/map_config.cc
@@ -1,1 +1,0 @@
-#include "sniff/map_config.h"
