--- src/minimize_config.cc
@@ -1,1 +1,0 @@
-#include "sniff/minimize_config.h"
